% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_estimate.R
\name{ComputeGaussianPL}
\alias{ComputeGaussianPL}
\title{Hyperparameter range.}
\usage{
ComputeGaussianPL(X, compute_distance = TRUE)
}
\arguments{
\item{X}{Cell coordinates matrix n x 2 or kernel matrix computed already}

\item{compute_distance}{Compute the distance matrix using generic function dist, default=TRUE}
}
\value{
This function returns the range for the hyperparameter in the Gaussian Kernel.
\item{lrang}{The range for the hyperparameter in the Gaussian kernel}
}
\description{
The range for the hyperparameter in the Gaussian Kernel, which was calculated from the coordinate of all spots.
}
