% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctcut1}
\alias{ctcut1}
\title{identify cell types in each domain}
\usage{
ctcut1(ctData, ilay, cutoff = 0, offset = NULL)
}
\arguments{
\item{ctData}{estimated coefficients for each cell type within each domain}

\item{ilay}{ith domain}

\item{cutoff}{the minimum percent of nonzero entries, default value is 0}

\item{offset}{the value add on the threshold to binarize coefficient from lasso regression, if not provide by user, then use the mean of all coefficients}
}
\value{
the cell type names that pass the filtering
}
\description{
identify cell types in each domain
}
