% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\docType{data}
\name{D2_data}
\alias{D2_data}
\title{D2_data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 1188 columns.
}
\usage{
D2_data
}
\description{
Read counts data in 2DPI group.
}
\keyword{data}
