% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\docType{data}
\name{D2_info}
\alias{D2_info}
\title{D2_info}
\format{
An object of class \code{data.frame} with 1188 rows and 2 columns.
}
\usage{
D2_info
}
\description{
Coordinate data in 2DPI group.
}
\keyword{data}
