% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\docType{data}
\name{D5_data}
\alias{D5_data}
\title{D5_data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 938 columns.
}
\usage{
D5_data
}
\description{
Read counts data in 5DPI group.
}
\keyword{data}
