% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\docType{data}
\name{D5_info}
\alias{D5_info}
\title{D5_info}
\format{
An object of class \code{data.frame} with 938 rows and 2 columns.
}
\usage{
D5_info
}
\description{
Coordinates data in 5DPI group.
}
\keyword{data}
