% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_DE.R
\name{Delta_fit}
\alias{Delta_fit}
\title{Estimation of the optimal delta}
\usage{
Delta_fit(location, y, L)
}
\arguments{
\item{location}{Spatial coordinates for all spots.}

\item{y}{Read counts data for the gene.}

\item{L}{The length-scale hyperparameter in the kernel function.}
}
\value{
This function returns the optimal log transformed delta value.
\item{results}{The optimal log delta value.}
}
\description{
Estimation of the optimal delta in the variance function between groups.
}
