% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dot_plot}
\alias{dot_plot}
\title{Dotplot to compare estimated proportion with true proportion}
\usage{
dot_plot(trueprop, estprop, nct, colors)
}
\arguments{
\item{trueprop}{true proportion}

\item{estprop}{estimated proportion}

\item{nct}{number of cell types}

\item{colors}{defind colors for each cell type}
}
\value{
dot plot
}
\description{
Dotplot to compare estimated proportion with true proportion
}
