% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{featureplot}
\alias{featureplot}
\title{scatter plot for cell types at each location}
\usage{
featureplot(propdat, locdat, celltype)
}
\arguments{
\item{propdat}{cell type proportion}

\item{locdat}{spatial location}

\item{celltype}{what cell types to plot}
}
\value{
scatterplot shows the cell type proportion at each location
}
\description{
scatter plot for cell types at each location
}
