% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\name{Get_Lambda}
\alias{Get_Lambda}
\title{Eigenvalue for Kernel matrix}
\usage{
Get_Lambda(K)
}
\arguments{
\item{K}{Kernel matrix.}
}
\value{
This function returns eigenvalue for the Kernel matrix.
\item{lambda}{Eigenvalue for Kernel matrix.}
}
\description{
Calculating the eigenvalue for the kernel matrix K.
}
