% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_estimate.R
\name{lengthscale_fit}
\alias{lengthscale_fit}
\title{Optimization of likelihood function within groups}
\usage{
lengthscale_fit(location, y, L)
}
\arguments{
\item{location}{Spatial coordinates for all spots.}

\item{y}{Read counts data for a gene.}

\item{L}{The length-scale hyperparameter in the kernel function.}
}
\value{
This function returns likelihood at the point of optimal delta.
\item{results}{The likelihood with the optimal delta value.}
}
\description{
Optimization of likelihood function within groups.
}
