% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_DE.R
\name{lengthscale_fit_DE}
\alias{lengthscale_fit_DE}
\title{Optimization of likelihood function between groups}
\usage{
lengthscale_fit_DE(location1, location2, y1, y2, L)
}
\arguments{
\item{location1}{Spatial coordinates for all spots in group 1.}

\item{location2}{Spatial coordinates for all spots in group 2.}

\item{y1}{Read counts data for the gene in group 1.}

\item{y2}{Read counts data for the corresponding gene in group 2.}

\item{L}{The length-scale hyperparameter in the kernel function.}
}
\value{
This function returns likelihood at the point of optimal delta.
\item{results}{The likelihood with the optimal delta value for identifying SE genes between groups.}
}
\description{
Optimization of likelihood function between groups.
}
