% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_estimate.R
\name{LL}
\alias{LL}
\title{Log likelihood function within groups}
\usage{
LL(log_delta, UTy, UT1, S, n)
}
\arguments{
\item{log_delta}{Log transformed delta in the variance function.}

\item{UTy}{The expression UTy will need to be re-computed for each gene.}

\item{UT1}{The expression UT1 only depends on the coordinates X and can be pre-computed and reused for every gene.}

\item{S}{A diagonal matrix used in the spectral decomposition of the covariance.}

\item{n}{The number of spots for each gene.}
}
\value{
This function returns the log likelihood for each delta.
\item{LL_res}{The log likelihood for each delta value.}
}
\description{
The log transformed multivariate normally distributed marginal likelihood to identify SE genes within groups.
}
