% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_DE.R
\name{LL_combin}
\alias{LL_combin}
\title{Log likelihood function between groups}
\usage{
LL_combin(log_delta, UTy1, UTy2, UT1, UT2, S1, S2, n1, n2)
}
\arguments{
\item{log_delta}{Log transformed delta in the variance function.}

\item{UTy1}{The expression UTy will need to be re-computed for the gene in group 1.}

\item{UTy2}{The expression UTy will need to be re-computed for the gene in group 2.}

\item{UT1}{The expression UT1 only depends on the coordinates X and can be pre-computed and reused for the gene in group 1.}

\item{UT2}{The expression UT1 only depends on the coordinates X and can be pre-computed and reused for the gene in group 2.}

\item{S1}{A diagonal matrix used in the spectral decomposition of the covariance in group 1.}

\item{S2}{A diagonal matrix used in the spectral decomposition of the covariance in group 2.}

\item{n1}{The number of spots for each gene in group 1.}

\item{n2}{The number of spots for each gene in group 2.}
}
\value{
This function returns the log likelihood for each delta.
\item{LL_res}{The log likelihood for each delta value.}
}
\description{
The log transformed multivariate normally distributed marginal likelihood to identify SE genes between groups.
}
