% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_DE.R
\name{LL_DE}
\alias{LL_DE}
\title{Log likelihood function}
\usage{
LL_DE(log_delta, location, y, L)
}
\arguments{
\item{log_delta}{Log transformed delta in the variance function.}

\item{location}{Spatial coordinates for all spots.}

\item{y}{Read counts data for the gene.}

\item{L}{The length-scale hyperparameter in the kernel function.}
}
\value{
This function returns the log likelihood for each delta.
\item{LL_res}{The log likelihood for each delta value.}
}
\description{
The log transformed multivariate normally distributed marginal likelihood .
}
