% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normcount}
\alias{normcount}
\title{Normalize count data}
\usage{
normcount(count, method)
}
\arguments{
\item{count}{raw data}

\item{method}{name which normalization method to use}
}
\value{
normalized data
}
\description{
Normalize count data
}
