% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_norm.R
\name{regress_out}
\alias{regress_out}
\title{SPADE residualize data}
\usage{
regress_out(
  sample_info,
  expression_matrix,
  covariate_formula,
  design_formula = "~ 1"
)
}
\arguments{
\item{sample_info}{Spatial coordinates for all spots.}

\item{expression_matrix}{Stabilized data from above step. Each row is a gene and each column is a spot.}

\item{covariate_formula}{Specify covariates to regress out.}

\item{design_formula}{Design formula.}
}
\value{
This function returns the residualize data.
\item{regressed}{The residualized data}
}
\description{
Implementation of limma's removeBatchEffect function to residualize data.
}
