% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{scatter_pie}
\alias{scatter_pie}
\title{Scatter pie plot for cell type composition at each location}
\usage{
scatter_pie(fuldat, dat, fillcolors)
}
\arguments{
\item{fuldat}{data.frame contains cell type proportion and spatial location}

\item{dat}{data.frame for cell type proportion}

\item{fillcolors}{identify colors for each cell type}
}
\value{
scatter pie plot
}
\description{
Scatter pie plot for cell type composition at each location
}
