% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{scRefer}
\alias{scRefer}
\title{construct cell type reference from scRNA-seq data}
\usage{
scRefer(input_data, ct_var, sample_var)
}
\arguments{
\item{input_data}{an expressionSet class contains single cell gene expression and its metadata which contains annoted cell type}

\item{ct_var}{variable name for cell type column}

\item{sample_var}{variable name for sample column}
}
\value{
a matrix contains gene expression for each cell type
}
\description{
construct cell type reference from scRNA-seq data
}
