% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\docType{data}
\name{SeqFISH}
\alias{SeqFISH}
\title{SeqFISH}
\format{
An object of class \code{data.frame} with 249 rows and 131 columns.
}
\usage{
SeqFISH
}
\description{
Read counts data of SeqFISH.
}
\keyword{data}
