\name{SPADE-package}
\alias{SPADE-package}
\docType{package}
\title{
Spatial Population Abundance Dynamics Engine
}
\description{
Runs deterministic mathematical models predicting population abundance over time,
incorporating multiple species and management strategies.
}
\details{
\tabular{ll}{
Package: \tab SPADE\cr
Type: \tab Package\cr
Version: \tab 0.0 alpha\cr
Date: \tab 2013-06-25\cr
License: \tab Unreleased\cr
}
}
\author{
Nick Beeton, Grant Williamson, Joanne Potts et al.


Maintainer: nick.beeton@utas.edu.au
}
\references{
McMahon et al. 2010
}
\keyword{ package }
\seealso{
}
\examples{
}
