\name{SPADE.flattenAnnotations}
\alias{SPADE.flattenAnnotations}
\title{
Flatten list of annotations to matrix
}
\description{
Helper function for flattening list of annotations 
}
\usage{
SPADE.flattenAnnotations(annotations)
}

\arguments{
  \item{annotations}{
A list of annotation matrices. All matrices must have the same number of rows.
}
}
\value{
Single matrix of annotations
}
\author{
Michael Linderman
}
