% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE.R
\name{SPADE}
\alias{SPADE}
\title{SPADE deconvolution for each spatial domain}
\usage{
SPADE(
  stcount,
  scref,
  sign_list,
  loc,
  ctData,
  SVG,
  cutoff = 0,
  offset,
  yNorm = "cpm",
  bNorm = "cpm"
)
}
\arguments{
\item{stcount}{spatial gene expression}

\item{scref}{cell type reference from single cell RNA-seq data}

\item{sign_list}{marker genes lists for each cell type}

\item{loc}{data.frame contains location information}

\item{ctData}{estimated coefficients for each cell type within each domain}

\item{SVG}{differentially expressed genes between domains}

\item{cutoff}{the minimum percent of nonzero entries, default value is 0}

\item{offset}{the value add on the threshold to binarize coefficient from lasso regression, if not provide by user, then use the mean of all coefficients}

\item{yNorm}{the normalization method for spatial gene expression, default is cpm}

\item{bNorm}{the normalization method for reference, default is cpm.}
}
\value{
a matrix for cell type proportion
}
\description{
SPADE deconvolution for each spatial domain
}
