% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_DE.R
\name{SPADE_DE}
\alias{SPADE_DE}
\title{SPADE estimation and test between groups}
\usage{
SPADE_DE(
  readcounts1,
  readcounts2,
  location1,
  location2,
  mode = "Shape&Strength"
)
}
\arguments{
\item{readcounts1}{Read counts data for identifying spatially expressed genes in group 1. Each row is a gene and each column is a spot.}

\item{readcounts2}{Read counts data for identifying spatially expressed genes in group 2. Each row is a gene and each column is a spot.}

\item{location1}{Spatial coordinates for all spots in group 1.}

\item{location2}{Spatial coordinates for all spots in group 2.}

\item{mode}{Different modes for testing including 1. "Strength" to compare pattern strength, 2. "Shape" to compare pattern shapes, 3. "Shape&Strength" to compare pattern shapes and strength together.}
}
\value{
This function returns the estimated parameters and some statistics of the SPADE method.
\item{geneid}{Gene index.}
\item{theta_Gau1}{The estimated optimal length-scale hyperparameter in the Gaussian kernel for group 1.}
\item{theta_Gau2}{The estimated optimal length-scale hyperparameter in the Gaussian kernel for group 2.}
\item{Gamma1}{The estimated optimal Gamma1 for group 1.}
\item{Gamma2}{The estimated optimal Gamma1 for group 2.}
\item{logLik11}{The corresponding log likelihood calculated with optimal hyperparameter estimated above for group 1.}
\item{logLik21}{The corresponding log likelihood calculated with optimal hyperparameter estimated above for group 2.}
\item{logLik10}{The log likelihood calculated with expression value from group 1 and optimal hyperparameter estimated for group 2.}
\item{logLik20}{The log likelihood calculated with expression value from group 2 and optimal hyperparameter estimated for group 1.}
\item{Diff}{The likelihood ratio test statistic is given by 2*(logLik11 + logLik21 - logLik10 - logLik20).}
\item{Pvalue}{P values calculated with likelihood ratio test statistic using F test with degree freedom of one.}
\item{Adjust.Pvalue}{Adjusted P values calculated with the Benjamini and Hochberg method.}
}
\description{
Estimation of optimal parameter in the kernel function to identify SE genes between groups. P values will be provided based on a crossed likelihood ratio test.
}
