% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_estimate.R
\name{SPADE_estimate}
\alias{SPADE_estimate}
\title{SPADE estimation for hyperparameter within groups}
\usage{
SPADE_estimate(expr_data, info)
}
\arguments{
\item{expr_data}{Read counts data for identifying spatially expressed genes. Each row is a gene and each column is a spot.}

\item{info}{Spatial coordinates for all spots.}
}
\value{
This function returns the estimated parameters and some statistics of the SPADE method.
\item{GeneID}{Gene index.}
\item{theta_Gau}{The estimated optimal length-scale hyperparameter in the Gaussian kernel}
\item{Gamma_hat}{The estimated optimal Gamma_1 in the covariance function.}
\item{Lik_Gau}{The log likelihood using optimal hyparameter estimated above.}
}
\description{
Estimation of optimal hyperparameter in the kernel function to identify SE genes within groups.
}
\examples{
Y <- matrix(rnorm(10000, 10, 2),100, 100)
info <- matrix(runif(200, 1, 100), 100, 2)
output <- SPADE_estimate(expr_data=Y, info=info)
output

}
