% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_norm.R
\name{SPADE_norm}
\alias{SPADE_norm}
\title{SPADE normalization}
\usage{
SPADE_norm(readcounts, info)
}
\arguments{
\item{readcounts}{Read counts data for identifying spatially expressed genes. Each row is a gene and each column is a spot.}

\item{info}{Spatial coordinates for all spots.}
}
\value{
This function returns the normalized continuous data.
\item{regdata}{The normlized continuous data}
}
\description{
Normlization of read counts data into continuous data.
}
