% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_test.R
\name{SPADE_test}
\alias{SPADE_test}
\title{SPADE test within groups}
\usage{
SPADE_test(object, location, para)
}
\arguments{
\item{object}{Read counts data for identifying spatially expressed genes. Each row is a gene and each column is a spot.}

\item{location}{Spatial coordinates for all spots.}

\item{para}{Parameter data frame estimated from the SPADE_estimate function.}
}
\value{
This function returns the statistics and P values for the test
\item{GeneID}{Gene index.}
\item{Q}{The Q statistics calculated using the SKAT method.}
\item{Pvalue}{P value calculated from the Q statistics based on the Davies method.}
\item{Adjust.Pvalue}{Adjusted P values calculated with the Benjamini and Hochberg method.}
}
\description{
Test for identifying SE genes within groups
}
