% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE.R
\name{spadeDecon}
\alias{spadeDecon}
\title{SPADE:constrained nonlinear optimization to estimate cell type proportion}
\usage{
spadeDecon(bulk, reference, w = 1, yNorm = "cpm", bNorm = "cpm")
}
\arguments{
\item{bulk}{spatial gene expression}

\item{reference}{cell type reference}

\item{w}{weights for genes, default is 1}

\item{yNorm}{normalization method for spatial data, default is cpm}

\item{bNorm}{normalization method for reference, default is cpm}
}
\description{
SPADE:constrained nonlinear optimization to estimate cell type proportion
}
