% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE_norm.R
\name{stabilize}
\alias{stabilize}
\title{SPADE stabilize variance}
\usage{
stabilize(expression_matrix)
}
\arguments{
\item{expression_matrix}{Read counts data for identifying spatially expressed genes. Each row is a gene and each column is a spot.}
}
\value{
This function returns the stabilized data.
\item{stabilized_matrix}{The stabilized data}
}
\description{
Use Anscombe's approximation to variance stabilize Negative Binomial data.
}
